move sum() and complementaryPercent() into mds-utils
