+++ packages/mds-metrics-sheet/tests/mds-metrics-sheet.test.ts
@@ -53,15 +53,6 @@ describe('MDS Metrics Sheet', () => {
       }
       assert.deepStrictEqual(result, expected)
     })
-
-    it('Computes `sum()` correctly', () => {
-      const arr = [1, 2, 3]
-      assert.equal(metricsLogUtils.sum(arr), 6)
-    })
-
-    it('Computes `complementaryPercent()` correctly', () => {
-      assert.equal(metricsLogUtils.complementaryPercent(9, 100), 0.91)
-    })
   })
 
   it('Maps empty row correctly', () => {
